refactor(api-url): 엔드포인트 상수 정의한 파일 이름 변경한다
