+++ src/lib/api/panel.ts
@@ -1,7 +1,7 @@
 import type { Member } from '@/lib/api/auth';
 import type { SuccessResponse } from '@/lib/api/types';
 
-import { apiUrl } from '@/lib/api/consts';
+import { apiUrl } from '@/lib/api/apiUrl';
 import { apiClient } from '@/lib/apiClient';
 
 export const createPanel = async (
