+++ src/mocks/handlers/auth.ts
@@ -12,7 +12,7 @@ import type { ErrorResponse } from '@/lib/api/types';
 
 import { rest } from 'msw';
 
-import { apiUrl } from '@/lib/api/consts';
+import { apiUrl } from '@/lib/api/apiUrl';
 import { API_BASE_URL } from '@/lib/apiClient';
 import { myAccount } from '@/mocks/data/auth';
