+++ src/pages/__tests__/panel-route.test.tsx
@@ -8,7 +8,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import { rest } from 'msw';
 import { RouterProvider, createMemoryRouter } from 'react-router-dom';
 
-import { apiUrl } from '@/lib/api/consts';
+import { apiUrl } from '@/lib/api/apiUrl';
 import * as panelApis from '@/lib/api/panel';
 import { server } from '@/mocks/server';
 import { Panel, panelLoader, PanelError } from '@/pages/Panel';
